requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
mysql-connector-python==8.1.0
numpy==1.25.1
pandas==2.0.3
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
style==1.1.0
tzdata==2023.3
update==0.0.1
urllib3==2.0.4
